Fix stuttering and high CPU during playback
The AVAudioPlayer does not like it if we schedule a lot of small
buffers on it. It soon starts maxing out the CPU, and audio playback
is reduced to stutters. So this commit was work towards buffer and
convert a second's worth of audio data at a time.

When I finally got it done, I figured out that it was
unnecessary. Something else was wrong in the last implementation, but
we do not necessarily need to buffer a whole second worth of data -
that buffering code is infact not working currently and commented, yet
we achieve smooth low CPU playback. Committing this WIPish code in
order to have a copy around in history; will trim out the unused bits
in a subsequent commit.